fully flesh out lint tests without protobuf-gradle-plugin
