+++ src/test/resources/LintTest/lint a file with a google dependency/build.gradle
