+++ src/test/resources/LintTest/lint with a config in default location/test.proto
