+++ src/test/resources/LintTest/lint with a dependency config override fails with two files/build.gradle
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Andrew Parmet
+ * Copyright (c) 2022 Andrew Parmet
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,19 +15,8 @@
 
 plugins {
   id 'java'
-  id 'com.google.protobuf' version "$protobufGradleVersion"
   id 'com.parmet.buf'
 }
-
-repositories { mavenCentral() }
-
-protobuf {
-  protoc {
-    artifact = "com.google.protobuf:protoc:$protobufVersion"
-  }
-}
-
-compileJava.enabled = false
 
 dependencies {
   buf(fileTree('subdir') { include '*.yaml' })
