+++ src/test/resources/LintTest/lint with a dependency config override/test.proto
