+++ src/test/resources/LintWithProtobufGradleTest/lint a basic correct message with default config/src/main/proto/parmet/buf/test/v1/test.proto
