+++ src/test/resources/LintWithProtobufGradleTest/lint a basic incorrect message with wrong location/src/main/proto/test.proto
