+++ src/test/resources/LintWithProtobufGradleTest/lint a file with a google dependency/build.gradle
