+++ src/test/resources/LintWithProtobufGradleTest/lint a file with a google dependency/src/main/proto/parmet/buf/test/v1/test.proto
