+++ src/test/resources/LintWithProtobufGradleTest/lint a file with a protobuf dependency and a google dependency with the protobuf-gradle-plugin/build.gradle
