+++ src/test/resources/LintWithProtobufGradleTest/lint a file with a protobuf dependency with the protobuf-gradle-plugin/src/main/proto/parmet/buf/test/v1/test.proto
