+++ src/test/resources/LintWithProtobufGradleTest/lint a separate protobuf source directory through the protobuf-gradle-plugin/build.gradle
