+++ src/test/resources/LintWithProtobufGradleTest/lint a separate protobuf source directory through the protobuf-gradle-plugin/subdir/parmet/buf/test/v1/test.proto
