+++ src/test/resources/LintWithProtobufGradleTest/lint with a config in default location/src/main/proto/parmet/buf/test/test.proto
