+++ src/test/resources/LintWithProtobufGradleTest/lint with a dependency config override fails with no files/src/main/proto/parmet/buf/test/v1/test.proto
