+++ src/test/resources/LintWithProtobufGradleTest/lint with a dependency config override fails with two files/subdir/buf-1.yaml
@@ -1,4 +1,4 @@
-# Copyright (c) 2022 Andrew Parmet
+# Copyright (c) 2021 Andrew Parmet
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
