+++ src/test/resources/LintWithProtobufGradleTest/lint with a dependency config override/src/main/proto/parmet/buf/test/test.proto
