+++ src/test/resources/LintWithProtobufGradleTest/lint with a file and dependency config override fails/src/main/proto/parmet/buf/test/v1/test.proto
