+++ src/test/resources/LintWithProtobufGradleTest/lint with a file location config override/src/main/proto/parmet/buf/test/test.proto
